「🔨」 fix(libft): fixed compilation errors

+++ libft/include/libft.h
@@ -6,7 +6,7 @@
 /*   By: mmoussou <mmoussou@student.42angouleme.fr  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/30 23:06:24 by mmoussou          #+#    #+#             */
-/*   Updated: 2024/04/29 13:29:47 by mmoussou         ###   ########.fr       */
+/*   Updated: 2024/04/30 11:03:55 by mmoussou         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -56,7 +56,7 @@ char		*ft_ltoa(long long n);
 
 /* STRINGS */
 
-size_t		ft_strlen(char const *s);
+uint		ft_strlen(char const *s);
 char		*ft_strchr(const char *s, int c);
 char		*ft_strrchr(const char *s, int c);
 int			ft_strcmp(const char *s1, const char *s2);
